LPS-60609 Move it back to make it consistent with other DDM modules where the exceptions are in the main folder
