+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-validator/src/main/java/com/liferay/dynamic/data/mapping/validator/internal/DDMFormLayoutValidatorImpl.java
@@ -19,13 +19,13 @@ import com.liferay.dynamic.data.mapping.model.DDMFormLayoutColumn;
 import com.liferay.dynamic.data.mapping.model.DDMFormLayoutPage;
 import com.liferay.dynamic.data.mapping.model.DDMFormLayoutRow;
 import com.liferay.dynamic.data.mapping.model.LocalizedValue;
+import com.liferay.dynamic.data.mapping.validator.DDMFormLayoutValidationException;
 import com.liferay.dynamic.data.mapping.validator.DDMFormLayoutValidator;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormLayoutValidationException;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormLayoutValidationException.DefaultLocaleNotSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormLayoutValidationException.DuplicateFieldName;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormLayoutValidationException.InvalidColumnSize;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormLayoutValidationException.InvalidDefaultLocaleSetForPageTitle;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormLayoutValidationException.InvalidRowSize;
+import com.liferay.dynamic.data.mapping.validator.DDMFormLayoutValidationException.DefaultLocaleNotSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormLayoutValidationException.DuplicateFieldName;
+import com.liferay.dynamic.data.mapping.validator.DDMFormLayoutValidationException.InvalidColumnSize;
+import com.liferay.dynamic.data.mapping.validator.DDMFormLayoutValidationException.InvalidDefaultLocaleSetForPageTitle;
+import com.liferay.dynamic.data.mapping.validator.DDMFormLayoutValidationException.InvalidRowSize;
 import com.liferay.portal.kernel.util.SetUtil;
 
 import java.util.HashSet;
