+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-validator/src/main/java/com/liferay/dynamic/data/mapping/validator/internal/DDMFormValidatorImpl.java
@@ -19,19 +19,19 @@ import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.model.DDMFormFieldOptions;
 import com.liferay.dynamic.data.mapping.model.DDMFormFieldType;
 import com.liferay.dynamic.data.mapping.model.LocalizedValue;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException;
 import com.liferay.dynamic.data.mapping.validator.DDMFormValidator;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.AvailableLocalesNotSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.DefaultLocaleNotSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.DefaultLocaleNotSetAsAvailableLocale;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.DuplicateFieldName;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.EmptyOptionSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.FieldTypeNotSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidAvailableLocalesSetForProperty;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidCharactersSetForFieldName;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidCharactersSetForFieldType;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidDefaultLocaleSetForProperty;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidIndexTypeSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.AvailableLocalesNotSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.DefaultLocaleNotSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.DefaultLocaleNotSetAsAvailableLocale;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.DuplicateFieldName;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.EmptyOptionSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.FieldTypeNotSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidAvailableLocalesSetForProperty;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidCharactersSetForFieldName;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidCharactersSetForFieldType;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidDefaultLocaleSetForProperty;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidIndexTypeSet;
 import com.liferay.portal.kernel.bean.BeanPropertiesUtil;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.MapUtil;
