+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-validator/src/test/java/com/liferay/dynamic/data/mapping/validator/DDMFormValidatorTest.java
@@ -23,17 +23,17 @@ import com.liferay.dynamic.data.mapping.model.DDMFormFieldOptions;
 import com.liferay.dynamic.data.mapping.model.DDMFormFieldType;
 import com.liferay.dynamic.data.mapping.model.LocalizedValue;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.AvailableLocalesNotSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.DefaultLocaleNotSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.DefaultLocaleNotSetAsAvailableLocale;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.DuplicateFieldName;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.EmptyOptionSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.FieldTypeNotSet;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidAvailableLocalesSetForProperty;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidCharactersSetForFieldName;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidCharactersSetForFieldType;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidDefaultLocaleSetForProperty;
-import com.liferay.dynamic.data.mapping.validator.exception.DDMFormValidationException.InvalidIndexTypeSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.AvailableLocalesNotSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.DefaultLocaleNotSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.DefaultLocaleNotSetAsAvailableLocale;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.DuplicateFieldName;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.EmptyOptionSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.FieldTypeNotSet;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidAvailableLocalesSetForProperty;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidCharactersSetForFieldName;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidCharactersSetForFieldType;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidDefaultLocaleSetForProperty;
+import com.liferay.dynamic.data.mapping.validator.DDMFormValidationException.InvalidIndexTypeSet;
 import com.liferay.dynamic.data.mapping.validator.internal.DDMFormValidatorImpl;
 import com.liferay.portal.bean.BeanPropertiesImpl;
 import com.liferay.portal.kernel.bean.BeanPropertiesUtil;
@@ -47,7 +47,6 @@ import java.util.Set;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.core.classloader.annotations.SuppressStaticInitializationFor;
